Moving modifiers and extensions related to UI to the UILibrary and create HomeView and HomeViewModel
