+++ modules/Feature/Package.swift
@@ -32,6 +32,7 @@ let package = Package(
                 .product(name: "ServicesInterface", package: "CoreLibrary"),
                 // UI
                 .product(name: "UIFoundations", package: "UILibrary"),
+                .product(name: "SwiftUIComponents", package: "UILibrary")
             ]
         ),
         .testTarget(
