+++ modules/UILibrary/Sources/SwiftUIComponents/Extension/Binding+WrapValue.swift
@@ -1,0 +1,22 @@
+import Foundation
+import SwiftUI
+
+public extension Binding where Value == Bool {
+    init<Wrapped>(bindingOptional: Binding<Wrapped?>) {
+        self.init(
+            get: {
+                bindingOptional.wrappedValue != nil
+            },
+            set: { newValue in
+                guard newValue == false else { return }
+                bindingOptional.wrappedValue = nil
+            }
+        )
+    }
+}
+
+extension Binding {
+    public func wrapValueToBool<Wrapped>() -> Binding<Bool> where Value == Wrapped? {
+        return Binding<Bool>(bindingOptional: self)
+    }
+}
