+++ modules/UILibrary/Sources/SwiftUIComponents/NavigationStackModifier.swift
@@ -1,0 +1,29 @@
+import Foundation
+import SwiftUI
+
+struct NavigationStackModifier<Item, Destination: View>: ViewModifier {
+    let item: Binding<Item?>
+    let destination: (Item) -> Destination
+    
+    func body(content: Content) -> some View {
+        // TODO: Apply to a NavigationStack and see how it works
+        content.background(NavigationLink(isActive: item.wrapValueToBool()) {
+            if let item = item.wrappedValue {
+                destination(item)
+            } else {
+                EmptyView()
+            }
+        } label: {
+            EmptyView()
+        })
+    }
+}
+
+public extension View {
+    func navigationDestination<Item, Destination: View>(
+        for binding: Binding<Item?>,
+        @ViewBuilder destination: @escaping (Item) -> Destination
+    ) -> some View {
+        self.modifier(NavigationStackModifier(item: binding, destination: destination))
+    }
+}
